qa: fix tests that fail under PHP 8
For `ErrorHandlerTest`, division by zero no longer results in an error, but a Throwable, which means we need to explicitly create an error.

For `ArrayObjectTest`, we need to change expectations based on PHP version, as what raised a warning before now raises a `TypeError`.

In both cases, the classes will work as expected; it is the underlying language behavior that changes how we test for the expected behavior.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>