feat: PHP 8.1 support via polyfills
When testing with `error_reporting = E_ALL`, we get deprecation notices that lead to fatal errors when a class implementing an internal interface differs in signature.
In order to resolve those issues, we need to either (a) correct the signature, or (b) add an attribute (e.g. `#[ReturnTypeWillChange]`).
Since both require either changing the code in a BC-impcompatible way, I am proposing we:

- Create a "legacy" and "PHP 8.1" implementation, with the "legacy" implementation being one unchanged from the current implementation, and the "PHP 8.1" implementation performing the requisite signature changes to allow the code to run under that version and above.
- Add an autoloader that performs a `class_alias`, aliasing the appropriate implementation to the original class name.

This approach allows tests to run successfully, and ensures the code runs correctly across versions.
The potential downside is that anybody updating to PHP 8.1 who is extending these classes AND overriding affected methods MAY experience BC breakage.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>