fix: fix issues with invalid imports due to usage of class aliases
One of our CS rules assumes that if a class isn't explicitly defined, it does not exist, and therefor references to it should be lowercased.
This causes invalid imports, and leads to odd code errors.
This patch adds rules to exclude that rule when using aliased classes to ensure the import statements are not rewritten.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>